20.04.19-1 session_log operation_log 完成

+++ src/main/java/com/yaqiu/constant/GlobalConstant.java
@@ -6,6 +6,6 @@ package com.yaqiu.constant;
  */
 public class GlobalConstant {
     /* [OPERATION_LOG表]的[TYPE]字段 */
-    public static final int PAGE_VISIT_OPERATION_LOG_TYPE = 0; //"页面访问"类型
-    public static final int SEND_REQUEST_OPERATION_LOG_TYPE = 1; //"发送请求"类型
+    public static final byte PAGE_VISIT_OPERATION_LOG_TYPE = 0; //"页面访问"类型
+    public static final byte SEND_REQUEST_OPERATION_LOG_TYPE = 1; //"发送请求"类型
 }
